# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Builder Pro API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectSourceType =
  (typeof ProjectSourceType)[keyof typeof ProjectSourceType];

export const ProjectSourceType = {
  github: "github",
  zip: "zip",
  local: "local",
} as const;

export type ProjectStatus = (typeof ProjectStatus)[keyof typeof ProjectStatus];

export const ProjectStatus = {
  pending: "pending",
  detecting: "detecting",
  ready: "ready",
  building: "building",
  failed: "failed",
} as const;

export interface Project {
  id: number;
  name: string;
  sourceType: ProjectSourceType;
  sourceUrl?: string | null;
  /** Detected type: react, vue, nextjs, angular, html, node */
  projectType?: string | null;
  packageId?: string | null;
  appName?: string | null;
  versionName?: string | null;
  versionCode?: number | null;
  status: ProjectStatus;
  createdAt: string;
  updatedAt: string;
}

export type NewProjectSourceType =
  (typeof NewProjectSourceType)[keyof typeof NewProjectSourceType];

export const NewProjectSourceType = {
  github: "github",
  zip: "zip",
  local: "local",
} as const;

export interface NewProject {
  name: string;
  sourceType: NewProjectSourceType;
  sourceUrl?: string | null;
  packageId?: string | null;
  appName?: string | null;
  versionName?: string | null;
  versionCode?: number | null;
}

export interface ProjectDetection {
  projectType: string;
  framework?: string | null;
  buildTool?: string | null;
  nodeVersion?: string | null;
  hasPackageJson: boolean;
  warnings: string[];
  recommendations: string[];
}

export type BuildConfigBuildType =
  (typeof BuildConfigBuildType)[keyof typeof BuildConfigBuildType];

export const BuildConfigBuildType = {
  debug: "debug",
  release: "release",
} as const;

export type BuildConfigOutputFormat =
  (typeof BuildConfigOutputFormat)[keyof typeof BuildConfigOutputFormat];

export const BuildConfigOutputFormat = {
  apk: "apk",
  aab: "aab",
} as const;

export interface BuildConfig {
  buildType: BuildConfigBuildType;
  outputFormat: BuildConfigOutputFormat;
  minSdkVersion?: number | null;
  targetSdkVersion?: number | null;
  useKeystore: boolean;
  /** Generate Play Store ready AAB with metadata */
  playStoreMode?: boolean;
  storeTitle?: string | null;
  storeDescription?: string | null;
  storeCategory?: string | null;
}

export type BuildBuildType =
  (typeof BuildBuildType)[keyof typeof BuildBuildType];

export const BuildBuildType = {
  debug: "debug",
  release: "release",
} as const;

export type BuildOutputFormat =
  (typeof BuildOutputFormat)[keyof typeof BuildOutputFormat];

export const BuildOutputFormat = {
  apk: "apk",
  aab: "aab",
} as const;

export type BuildStatus = (typeof BuildStatus)[keyof typeof BuildStatus];

export const BuildStatus = {
  queued: "queued",
  running: "running",
  success: "success",
  failed: "failed",
  cancelled: "cancelled",
} as const;

export interface Build {
  id: number;
  projectId: number;
  buildType: BuildBuildType;
  outputFormat: BuildOutputFormat;
  status: BuildStatus;
  /** 0-100 */
  progress: number;
  logs: string[];
  outputPath?: string | null;
  outputSizeBytes?: number | null;
  errorMessage?: string | null;
  playStoreMode: boolean;
  startedAt?: string | null;
  completedAt?: string | null;
  createdAt: string;
}

export interface Keystore {
  id: number;
  alias: string;
  commonName: string;
  organization?: string | null;
  validityYears: number;
  createdAt: string;
  expiresAt: string;
}

export interface NewKeystore {
  alias: string;
  password: string;
  commonName: string;
  organization?: string | null;
  country?: string | null;
  validityYears: number;
}

export interface DashboardStats {
  totalProjects: number;
  totalBuilds: number;
  successfulBuilds: number;
  failedBuilds: number;
  runningBuilds: number;
  totalOutputSizeMB: number;
  recentBuilds: Build[];
}

export interface SystemCheckItem {
  name: string;
  available: boolean;
  version?: string | null;
  required: boolean;
  installHint?: string | null;
}

export interface SystemCheck {
  allRequired: boolean;
  items: SystemCheckItem[];
}