# lib/api-zod/src/generated/types/build.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Builder Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { BuildBuildType } from "./buildBuildType";
import type { BuildOutputFormat } from "./buildOutputFormat";
import type { BuildStatus } from "./buildStatus";

export interface Build {
  id: number;
  projectId: number;
  buildType: BuildBuildType;
  outputFormat: BuildOutputFormat;
  status: BuildStatus;
  /** 0-100 */
  progress: number;
  logs: string[];
  outputPath?: string | null;
  outputSizeBytes?: number | null;
  errorMessage?: string | null;
  playStoreMode: boolean;
  startedAt?: string | null;
  completedAt?: string | null;
  createdAt: string;
}