# lib/api-zod/src/generated/types/buildConfig.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Builder Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { BuildConfigBuildType } from "./buildConfigBuildType";
import type { BuildConfigOutputFormat } from "./buildConfigOutputFormat";

export interface BuildConfig {
  buildType: BuildConfigBuildType;
  outputFormat: BuildConfigOutputFormat;
  minSdkVersion?: number | null;
  targetSdkVersion?: number | null;
  useKeystore: boolean;
  /** Generate Play Store ready AAB with metadata */
  playStoreMode?: boolean;
  storeTitle?: string | null;
  storeDescription?: string | null;
  storeCategory?: string | null;
}