# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Builder Pro API
 * OpenAPI spec version: 0.1.0
 */
import type { Build } from "./build";

export interface DashboardStats {
  totalProjects: number;
  totalBuilds: number;
  successfulBuilds: number;
  failedBuilds: number;
  runningBuilds: number;
  totalOutputSizeMB: number;
  recentBuilds: Build[];
}