# lib/api-zod/src/generated/types/keystore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Builder Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface Keystore {
  id: number;
  alias: string;
  commonName: string;
  organization?: string | null;
  validityYears: number;
  createdAt: string;
  expiresAt: string;
}