# lib/api-zod/src/generated/types/newKeystore.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Builder Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface NewKeystore {
  alias: string;
  password: string;
  commonName: string;
  organization?: string | null;
  country?: string | null;
  validityYears: number;
}