# lib/api-zod/src/generated/types/projectDetection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Builder Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectDetection {
  projectType: string;
  framework?: string | null;
  buildTool?: string | null;
  nodeVersion?: string | null;
  hasPackageJson: boolean;
  warnings: string[];
  recommendations: string[];
}