# lib/api-zod/src/generated/types/systemCheckItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * APK Builder Pro API
 * OpenAPI spec version: 0.1.0
 */

export interface SystemCheckItem {
  name: string;
  available: boolean;
  version?: string | null;
  required: boolean;
  installHint?: string | null;
}